refac: Extract feedback form behavior to own module
